Add typings, ignore mocha warnings, move integration tests to subdir
