+++ package.json
@@ -6,10 +6,12 @@
   "author": "rongyi.liu@ele.me",
   "license": "MIT",
   "scripts": {
-    "test": "mocha tests/index.js && cd tests && ./test.sh"
+    "test": "mocha tests && cd tests/integration && ./test.sh"
   },
   "types": "./types/index.d.ts",
   "devDependencies": {
+    "@types/expect": "^1.20.1",
+    "@types/mocha": "^2.2.41",
     "@types/node": "^7.0.13",
     "eslint": "^3.19.0",
     "eslint-config-elemefe": "^0.3.0",
