+++ tests/.eslintrc.yaml
@@ -1,0 +1,8 @@
+---
+  env:
+    es6: true
+    node: true
+    mocha: true
+  rules:
+    no-throw-literal: 0
+  extends: elemefe
