+++ tests/integration/invalid-number.thrift
